+++ lib/src/l10n.dart
@@ -6,13 +6,16 @@ SDUIL10n sduiL10 = const SDUIL10nEn();
 
 @immutable
 abstract class SDUIL10n {
-  const SDUIL10n(
-      {required this.error,
-      required this.loading,
-      required this.validationMalformedEmail,
-      required this.validationMalformedURL,
-      required this.validationMissingField,
-      required this.validationInvalidNumber});
+  const SDUIL10n({
+    required this.error,
+    required this.loading,
+    required this.validationMalformedEmail,
+    required this.validationMalformedURL,
+    required this.validationMissingField,
+    required this.validationInvalidNumber,
+    required this.selectDate,
+    required this.selectTime,
+  });
 
   final String error;
   final String loading;
@@ -20,6 +23,8 @@ abstract class SDUIL10n {
   final String validationMalformedEmail;
   final String validationMalformedURL;
   final String validationMissingField;
+  final String selectDate;
+  final String selectTime;
 }
 
 @immutable
@@ -32,6 +37,8 @@ class SDUIL10nEn extends SDUIL10n {
           validationMalformedEmail: 'The email address in not valid',
           validationMalformedURL: 'The URL in not valid',
           validationMissingField: 'The field is missing',
+          selectDate: 'Select date',
+          selectTime: 'Select time',
         );
 }
 
@@ -45,5 +52,7 @@ class SDUIL10nFr extends SDUIL10n {
           validationMalformedEmail: "L'adresse e-mail n'est pas valide",
           validationMalformedURL: "L'URL n'est pas valide",
           validationMissingField: 'Le champ est manquant',
+          selectDate: 'Selectionnez la date',
+          selectTime: 'Selectionnez l' 'heure',
         );
 }